LayoutTests/media/modern-media-controls/icon-button/icon-button-active-state.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing an <code>IconButton</code> has a white background-color when pressed.");

window.jsTestIsAsync = true;

const iconButton = new IconButton({ layoutDelegate: { layoutTraits: LayoutTraits.macOS } });
iconButton.iconName = Icons.Pause;
document.body.appendChild(iconButton.element);

let x, y;

function checkBackgroundColor()
{
    shouldBeEqualToString("window.getComputedStyle(iconButton.element).backgroundColor", "rgb(255, 255, 255)")
    if ("createTouch" in document) {
        testRunner.runUIScript(`
            uiController.liftUpAtPoint(${x}, ${y}, 1, function() {
                uiController.uiScriptComplete("Done");
            })`, finishJSTest);
    } else {
        eventSender.mouseUp();
        finishJSTest();
    }
}

scheduler.frameDidFire = function()
{
    const bounds = iconButton.element.getBoundingClientRect();
    if (bounds.height === 0)
        return;

    scheduler.frameDidFire = null;

    if ("testRunner" in window) {
        x = bounds.left + 1;
        y = bounds.top + 1;
        if ("createTouch" in document) {
            testRunner.runUIScript(`
                uiController.touchDownAtPoint(${x}, ${y}, 1, function() {
                    uiController.uiScriptComplete("Done");
                })`, checkBackgroundColor);
        } else {
            eventSender.mouseMoveTo(x, y);
            eventSender.mouseDown();
            checkBackgroundColor();
        }
    } else
        debug("This test is designed to run in DRT.");
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
